Add a system variable to exclude tables from the strict collation check
Summary: Some tables have case insensitive keys.  Normally these would be rejected when strict collation checking is enabled.  We need to provide a way to whitelist them as okay.  The new variable is rocksdb_strict_collation_exceptions and allows a list of tables (separated by commas and using regex) that will be allowed even when strict collation is turned on.

Test Plan: MTR

Reviewers: maykov, yoshinorim, spetrunia, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D49035
Differential Revision: https://reviews.facebook.net/D49215
Differential Revision: https://reviews.facebook.net/D49299
Differential Revision: https://reviews.facebook.net/D49323