Enable Howtoeven by default for everyone - brace for impact!
Summary: There's no good mode for selective opt-in and I really don't want to start polluting everyone's `.bashrc` or `.bash_profile`. Therefore we'll just be bold and enable Howtoeven for everyone. The worst-case scenario of plug-in blocking someone's workflow can be mitigated by setting `SKIP_HOWTOEVEN` to `1`. We'll keep that check in place for a couple of weeks (?) till we have validated the reliability with the sufficient data set and then we'll remove it.

Test Plan:
  # Make changes which cause Howtoeven to issue warnings.
  # Verify that correct warnings are issues and linter is executed.
  # Verify that after setting `export SKIP_HOWTOEVEN=1` the linter is not executed.

Reviewers: jkedgar

Reviewed By: jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D56085