Deprecate functionality with blind write batches and unique checks
Summary:
Disabling unique key checks adds keys to the blind write batch, but this
prevents the transaction from being able to see the key again. It was
originally using the blind write batch for performance reasons with bulk
loading, but now that bulk loading goes directly into the sst files, we
can use the index write batch for transactions that skip unique checks.

Reviewed By: alxyang

Differential Revision: D4946447

fbshipit-source-id: 7535209